Started work on windows support
- All code compiles on windows, needs testing
- Enabling/disabling image formats is now fully supported

+++ src/graphics/font.h
@@ -6,7 +6,7 @@
 #include "texture.h"
 
 typedef struct glyph {
-    int index;
+    int16 index;
     int16 texture_index;
     vec2 bearing;
     float advance;
